Make constructor arguments and fix role bug
By forcing the super class to have its properties passed by constructor, I can have tsc complain when I don't include all properties that aren't declared `someVar!: someType`, which makes it less prone to bugs when oops I forgot to instantiate with a certain property in the subclass.

Fix for the role command needed because the `Helper.getRoleByName` throws so that I don't need to constantly check for inside the commands. Since the loop needs to check if an argument is a possible role, it should ignore thrown errors in this case.